doubiodubi

+++ cyj-data-dictionary/src/main/java/com/example/cyjdatadictionary/dao/DictionaryCatalogDao.java
@@ -3,6 +3,6 @@ package com.example.cyjdatadictionary.dao;
 import com.example.cyjdatadictionary.entity.DictionaryCatalog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DictionaryCatalogDao extends JpaRepository<DictionaryCatalog,Long> {
-    DictionaryCatalog findDictionaryCatalogById(long id);
+public interface DictionaryCatalogDao extends JpaRepository<DictionaryCatalog,Integer> {
+    DictionaryCatalog findDictionaryCatalogById(Integer id);
 }
