+++ cyj-user/src/main/java/com/example/cyjuser/dao/UserDao.java
@@ -3,4 +3,4 @@ package com.example.cyjuser.dao;
 import com.example.cyjuser.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDao extends JpaRepository<User,Long> {}
+public interface UserDao extends JpaRepository<User,Integer> {}
